requirements.txt
fastapi
uvicorn[standard]
biopython
streamlit
requests
pandas
plotly